Add files via upload
All files currently being worked on. One thing I have learned is proper file set up. As this is a graded submission and progress in learning needs to be evident, I shall leave as is and correct after submission along with code clean up and removal of unnecessary comments and commented out code.